Refactoring and value added :squirrel:
