+++ package.json
@@ -42,12 +42,10 @@
     "vue-draggable-resizable": "^1.7.1",
     "vue-perfect-scrollbar": "^0.1.0",
     "vue-property-decorator": "^7.0.0",
-    "vue-svg-pan-zoom": "^0.1.0",
     "vuejs-logger": "^1.5.3",
     "vuetify": "^1.5.5",
     "vuex": "^3.0.1",
-    "vuex-module-decorators": "^0.9.7",
-    "wavesurfer.js": "^2.1.2"
+    "vuex-module-decorators": "^0.9.7"
   },
   "devDependencies": {
     "@types/chai": "^4.1.7",
