+++ src/background.ts
@@ -71,7 +71,7 @@ app.on('ready', async () => {
   if (isDevelopment && !process.env.IS_TEST) {
     // Install Vue Devtools
     try {
-      // await installVueDevtools();
+      await installVueDevtools();
     } catch (e) {
       // tslint:disable-next-line:no-console
       console.error('Vue Devtools failed to install:', e.toString());
