+++ src/modules/dawg/index.ts
@@ -6,8 +6,8 @@ import Dawg from '@/modules/dawg/Dawg.vue';
 import VerticalSwitch from '@/modules/dawg/VerticalSwitch.vue';
 import Blank from '@/modules/dawg/Blank.vue';
 import Loading from '@/modules/dawg/Loading.vue';
-import WaveformV2 from '@/modules/dawg/WaveformV2.vue';
-import SpectrogramV2 from '@/modules/dawg/SpectrogramV2.vue';
+import Waveform from '@/modules/dawg/Waveform.vue';
+import Spectrogram from '@/modules/dawg/Spectrogram.vue';
 
 export default  {
   install() {
@@ -18,8 +18,8 @@ export default  {
     Vue.component('VerticalSwitch', VerticalSwitch);
     Vue.component('Blank', Blank);
     Vue.component('Loading', Loading);
-    Vue.component('WaveformV2', WaveformV2);
-    Vue.component('SpectrogramV2', SpectrogramV2);
+    Vue.component('Waveform', Waveform);
+    Vue.component('Spectrogram', Spectrogram);
   },
 };
